feat: complete v1.2 remaining tasks and fix backend service startup issues
Backend fixes:
- TECH-OBS: V7 Flyway 'window' reserved word (quote column) + SloService @Qualifier on constructor
- TECH-MSG: V4 Flyway DOUBLE type -> DOUBLE PRECISION
- TECH-AGENT: Python 3.14 asyncio.get_event_loop() removal -> new_event_loop()
- TECH-ACTION: fix /definitions 500 (Hibernate 6 JPQL CONCAT bytea bug when keyword is null)
- TECH-GW: fix GrayReleaseFilter JSONB query (same JPQL bug)
- TECH-EA: switch H2 -> PostgreSQL for jsonb Flyway migrations

v1.2 frontend tasks:
- V12-02 APP-SUPERAI: GeneratePanel add "apply to designer" button + designerImport util
- V12-03 APP-APPHUB: integrate AI in Form/Flow/Page designers (consume localStorage import)
- V12-04 APP-APPHUB: page designer enhancements (custom component registry, page scripts, DataSourceBinding + useDataSource hook, 5 new widget types)
- V12-06 APP-ONTSTUDIO: DecisionTableEditor (matrix editor + 5 hit policies + 9 operators) + TestCaseManager (CRUD + batch run + diff) + TestRunner history
- V12-09 APP-DW: AutoScorePanel + OptimizationSuggestionsCard + DimensionScoreChart (SVG radar) + 4-tab EvaluationPage + 6-dimension scoring

Cleanup:
- Remove tracked pyc files (already gitignored)
- Remove temp debug files (TECH-IAM/reset-password.py, test-*.json)
- Add vite-env.d.ts to APP-SUPERAI and APP-DW for import.meta.env types

+++ TECH-MSG/src/main/resources/db/migration/V4__init_dlq_retry_policy_schema.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS msg_dlq_retry_policies (
     topic                    VARCHAR(256)   NOT NULL,
     max_retries              INTEGER        NOT NULL DEFAULT 3,
     retry_interval_seconds   INTEGER        NOT NULL DEFAULT 60,
-    retry_backoff_multiplier DOUBLE         NOT NULL DEFAULT 2.0,
+    retry_backoff_multiplier DOUBLE PRECISION NOT NULL DEFAULT 2.0,
     auto_cleanup_days        INTEGER        NOT NULL DEFAULT 30,
     created_at               TIMESTAMP      NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at               TIMESTAMP      NOT NULL DEFAULT CURRENT_TIMESTAMP,
